refactor: modularize browse helpers
